+++ syntheseus/reaction_prediction/inference/mhnreact.py
@@ -9,11 +9,11 @@ The original MHNreact code is released under the BSD-2-Clause license.
 import json
 from collections import defaultdict
 from functools import partial
-from typing import List
+from typing import List, Sequence
 
 from tqdm.contrib import concurrent
 
-from syntheseus.interface.models import BackwardPredictionList
+from syntheseus.interface.models import BackwardPrediction
 from syntheseus.interface.molecule import Molecule
 from syntheseus.reaction_prediction.inference.base import ExternalBackwardReactionModel
 from syntheseus.reaction_prediction.utils.inference import (
@@ -47,6 +47,8 @@ class MHNreactModel(ExternalBackwardReactionModel):
 
         with open(get_unique_file_in_dir(self.model_dir, pattern="*.json"), "r") as conf:
             conf_dict = json.load(conf)
+        conf_dict["device"] = self.device
+
         conf = model.ModelConfig(**conf_dict)
         self.model = model.MHN(config=conf)
 
@@ -93,7 +95,9 @@ class MHNreactModel(ExternalBackwardReactionModel):
     def get_parameters(self):
         return self.model.parameters()
 
-    def __call__(self, inputs: List[Molecule], num_results: int) -> List[BackwardPredictionList]:
+    def __call__(
+        self, inputs: List[Molecule], num_results: int
+    ) -> List[Sequence[BackwardPrediction]]:
         import pandas as pd
         import torch
 
@@ -151,7 +155,7 @@ class MHNreactModel(ExternalBackwardReactionModel):
         # Now aggregate over same outcome (parts copied from `utils.sort_by_template_and_flatten`,
         # which does not expose the summed probabilities) and build the prediction objects.
 
-        batch_predictions: List[BackwardPredictionList] = []
+        batch_predictions: List[Sequence[BackwardPrediction]] = []
         for idx in range(len(template_scores)):
             idx_prod_reactants = defaultdict(list)
             for k, v in prod_idx_reactants[idx].items():