+++ syntheseus/reaction_prediction/inference/retro_knn.py
@@ -4,11 +4,11 @@ Paper: https://arxiv.org/abs/2306.04123
 """
 
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import List, Optional, Sequence, Union
 
 import numpy as np
 
-from syntheseus.interface.models import BackwardPredictionList
+from syntheseus.interface.models import BackwardPrediction
 from syntheseus.interface.molecule import Molecule
 from syntheseus.reaction_prediction.inference.local_retro import LocalRetroModel
 from syntheseus.reaction_prediction.utils.inference import get_unique_file_in_dir
@@ -38,27 +38,31 @@ class RetroKNNModel(LocalRetroModel):
         import faiss
         import faiss.contrib.torch_utils  # make faiss available for torch tensors
 
-        def load_data_store(path: Path):
+        def load_data_store(path: Path, device: str):
             index = faiss.read_index(str(path), faiss.IO_FLAG_ONDISK_SAME_DIR)
-            res = faiss.StandardGpuResources()
-            co = faiss.GpuClonerOptions()
-            co.useFloat16 = True
-            return faiss.index_cpu_to_gpu(res, 0, index, co)
 
-        self.atom_store = load_data_store(datastore_path / "data.atom_idx")
-        self.bond_store = load_data_store(datastore_path / "data.bond_idx")
+            if device == "cpu":
+                return index
+            else:
+                res = faiss.StandardGpuResources()
+                co = faiss.GpuClonerOptions()
+                co.useFloat16 = True
+                return faiss.index_cpu_to_gpu(res, 0, index, co)
+
+        self.atom_store = load_data_store(datastore_path / "data.atom_idx", device=self.device)
+        self.bond_store = load_data_store(datastore_path / "data.bond_idx", device=self.device)
         self.raw_data = np.load(datastore_path / "data.npz")
 
-        self.adapter = Adapter(self.model.linearB.weight.shape[0], k=32).to(self.args["device"])
-        self.adapter.load_state_dict(torch.load(adapter_chkpt_path))
+        self.adapter = Adapter(self.model.linearB.weight.shape[0], k=32).to(self.device)
+        self.adapter.load_state_dict(torch.load(adapter_chkpt_path, map_location=self.device))
         self.adapter.eval()
 
     def _forward_localretro(self, bg):
         from local_retro.scripts.model_utils import pair_atom_feats, unbatch_feats, unbatch_mask
 
-        bg = bg.to(self.args["device"])
-        node_feats = bg.ndata.pop("h").to(self.args["device"])
-        edge_feats = bg.edata.pop("e").to(self.args["device"])
+        bg = bg.to(self.device)
+        node_feats = bg.ndata.pop("h").to(self.device)
+        edge_feats = bg.edata.pop("e").to(self.device)
 
         node_feats = self.model.mpnn(bg, node_feats, edge_feats)
         atom_feats = node_feats
@@ -72,7 +76,9 @@ class RetroKNNModel(LocalRetroModel):
 
         return atom_outs, bond_outs, atom_feats, bond_feats
 
-    def __call__(self, inputs: List[Molecule], num_results: int) -> List[BackwardPredictionList]:
+    def __call__(
+        self, inputs: List[Molecule], num_results: int
+    ) -> List[Sequence[BackwardPrediction]]:
         import torch
 
         from syntheseus.reaction_prediction.models.retro_knn import knn_prob
@@ -84,7 +90,7 @@ class RetroKNNModel(LocalRetroModel):
             atom_feats,
             bond_feats,
         ) = self._forward_localretro(batch)
-        sg = batch.remove_self_loop().to(self.args["device"])
+        sg = batch.remove_self_loop().to(self.device)
 
         node_dis, _ = self.atom_store.search(atom_feats, k=32)
         edge_dis, _ = self.bond_store.search(bond_feats, k=32)
@@ -96,12 +102,8 @@ class RetroKNNModel(LocalRetroModel):
         batch_atom_prob_nn = torch.nn.Softmax(dim=1)(batch_atom_logits)
         batch_bond_prob_nn = torch.nn.Softmax(dim=1)(batch_bond_logits)
 
-        atom_output_label = torch.from_numpy(self.raw_data["atom_output_label"]).to(
-            self.args["device"]
-        )
-        bond_output_label = torch.from_numpy(self.raw_data["bond_output_label"]).to(
-            self.args["device"]
-        )
+        atom_output_label = torch.from_numpy(self.raw_data["atom_output_label"]).to(self.device)
+        bond_output_label = torch.from_numpy(self.raw_data["bond_output_label"]).to(self.device)
 
         batch_atom_prob_knn = knn_prob(
             atom_feats, self.atom_store, atom_output_label, batch_atom_logits.shape[1], 32, node_t