+++ syntheseus/tests/reaction_prediction/inference/test_models.py
@@ -2,7 +2,15 @@ import pytest
 
 from syntheseus.interface.bag import Bag
 from syntheseus.interface.molecule import Molecule
-from syntheseus.reaction_prediction.inference import ChemformerModel, LocalRetroModel, MEGANModel
+from syntheseus.reaction_prediction.inference import (
+    ChemformerModel,
+    Graph2EditsModel,
+    LocalRetroModel,
+    MEGANModel,
+    MHNreactModel,
+    RetroKNNModel,
+    RootAlignedModel,
+)
 from syntheseus.reaction_prediction.inference.base import ExternalBackwardReactionModel
 
 try:
@@ -11,8 +19,11 @@ try:
     # the tests will fail; nevertheless the check below is good enough for our usecase.
 
     import chemformer  # noqa: F401
+    import graph2edits  # noqa: F401
     import local_retro  # noqa: F401
     import megan  # noqa: F401
+    import mhnreact  # noqa: F401
+    import root_aligned  # noqa: F401
 
     MODELS_INSTALLED = True
 except ModuleNotFoundError:
@@ -24,13 +35,18 @@ pytestmark = pytest.mark.skipif(
 )
 
 
-# TODO(kmaziarz): Some of the models (MHNreact, RetroKNN and RootAligned) appear to only work on
-# GPU. Make them also work on CPU, then add below.
-
-
 @pytest.fixture(
     scope="module",
-    params=[ChemformerModel, LocalRetroModel, MEGANModel],
+    params=[
+        ChemformerModel,
+        Graph2EditsModel,
+        LocalRetroModel,
+        MEGANModel,
+        MHNreactModel,
+        RetroKNNModel,
+        RootAlignedModel,
+    ]
+    * 2,
 )
 def model(request) -> ExternalBackwardReactionModel:
     model_cls = request.param
@@ -38,8 +54,8 @@ def model(request) -> ExternalBackwardReactionModel:
 
 
 def test_call(model: ExternalBackwardReactionModel) -> None:
-    [result] = model([Molecule("Cc1ccc(-c2ccc(C)cc2)cc1")], num_results=10)
-    model_predictions = [prediction.output for prediction in result.predictions]
+    [result] = model([Molecule("Cc1ccc(-c2ccc(C)cc2)cc1")], num_results=20)
+    model_predictions = [prediction.output for prediction in result]
 
     # Prepare some coupling reactions that are reasonable predictions for the product above.
     expected_predictions = [
@@ -48,7 +64,7 @@ def test_call(model: ExternalBackwardReactionModel) -> None:
         for leaving_group_2 in ["B(O)O", "I", "[Mg+]"]
     ]
 
-    # The model should recover at least two (out of six) in its top-10.
+    # The model should recover at least two (out of six) in its top-20.
     assert len(set(expected_predictions) & set(model_predictions)) >= 2